pyproject.toml
[tool.poetry]
name = "airflow-instance"
version = "0.1.0"
description = "airflow dags using poetry"
authors = ["Eric Butera <eric@ericbutera.com>"]

[tool.poetry.dependencies]
python = "^3.8,<3.11"
flake8 = "^4.0.1"
apache-airflow = "^2.3.4"
apache-airflow-providers-google = "^8.3.0"
pylint-airflow = "^0.1.0-alpha.1"
pre-commit = "^2.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pre-commit = "^2.19.0"
black = "^22.3.0"
flake8-black = "^0.3.3"
isort = "^5.10.1"
mypy = "^0.960"
pytest-mock = "^3.7.0"
pytest-datadir = "^1.3.1"
pytest-cov = "^3.0.0"
mock = "^4.0.3"
apache-airflow-providers-docker = "^3.1.0"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
